+++ rts/motoko-rts-tests/src/gc/utils.rs
@@ -18,10 +18,10 @@ pub const MAX_MARK_STACK_SIZE: usize = 100;
 pub enum GC {
     Copying,
     MarkCompact,
-    Experimental,
+    Generational,
 }
 
-pub static GC_IMPLS: [GC; 3] = [GC::Copying, GC::MarkCompact, GC::Experimental];
+pub static GC_IMPLS: [GC; 3] = [GC::Copying, GC::MarkCompact, GC::Generational];
 
 /// Read a little-endian (Wasm) word from given offset
 pub fn read_word(heap: &[u8], offset: usize) -> u32 {
