+++ src/exes/moc.ml
@@ -138,9 +138,9 @@ let argspec = [
     set_mode Compile ()), (* similar to --idl *)
       " compile and emit signature of stable types to `.most` file";
  
-  "--experimental-gc",
-  Arg.Unit (fun () -> Flags.gc_strategy := Mo_config.Flags.Experimental),
-  " use experimental GC";
+  "--generational-gc",
+  Arg.Unit (fun () -> Flags.gc_strategy := Mo_config.Flags.Generational),
+  " use generational GC";
 
   "--no-gc",
   Arg.Unit (fun () -> Flags.gc_strategy := Mo_config.Flags.No),
